+++ package.json
@@ -14,9 +14,6 @@
     "@react-pdf-viewer/core": "^3.12.0",
     "@react-pdf-viewer/default-layout": "^3.12.0",
     "@tailwindcss/postcss": "^4.1.3",
-    "next": "15.0.3",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
     "react-icons": "^5.3.0",
     "react-markdown": "^10.1.0",
     "react-pdf": "^9.2.1",
@@ -26,7 +23,11 @@
     "remark-html": "^13.0.2",
     "rsuite": "^5.74.2",
     "xml2js": "^0.6.2",
-    "zod": "^3.24.1"
+    "zod": "^3.24.1",
+    "next": "15.2.4",
+    "postcss": "^8.5.3",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0"
   },
   "devDependencies": {
     "@babel/preset-env": "^7.26.9",
@@ -35,6 +36,8 @@
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.3.0",
     "@types/jest": "^29.5.14",
+    "@eslint/eslintrc": "^3",
+    "@tailwindcss/postcss": "^4.1.3",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
@@ -46,7 +49,7 @@
     "jest-environment-jsdom": "^29.7.0",
     "postcss": "^8.5.3",
     "react-testing-library": "^8.0.1",
-    "tailwindcss": "^3.4.17",
+    "tailwindcss": "^4.1.3",
     "ts-jest": "^29.3.1",
     "ts-node": "^10.9.2",
     "typescript": "^5.8.3"
