--- post.config.mjs
@@ -1,6 +1,0 @@
-const config = {
-    plugins: {
-        "@tailwindcss/postcss": {},
-    },
-};
-export default config;
