--- tailwind.config.js
@@ -1,13 +1,0 @@
-// tailwind.config.js
-module.exports = {
-    theme: {
-      extend: {
-        colors: {
-          brandYellow: '#f9c000',
-          brandBlack: '#343a40',
-        },
-      },
-    },
-    plugins: [],
-  };
-
